Update forked TextInputLineComponent
